--- app/utils/nonce.ts
@@ -1,14 +1,0 @@
-import crypto from "crypto";
-
-export function nonce(): string {
-  const length = 15;
-  const bytes = crypto.randomBytes(length);
-
-  const nonce = bytes
-    .map((byte) => {
-      return byte % 10;
-    })
-    .join("");
-
-  return nonce;
-}
