+++ app/utils/session.server.ts
@@ -1,5 +1,4 @@
-import { nonce } from "./nonce";
-import crypto from "crypto";
+import { nonce, passesSecurityCheck } from "./security";
 import { createCookieSessionStorage, redirect } from "remix";
 
 const sessionSecret = process.env.SESSION_SECRET;
@@ -106,28 +105,4 @@ export async function handleCallback(request: Request) {
   } catch (error) {
     throw error;
   }
-}
-
-async function passesSecurityCheck(request: Request) {
-  const session = await getUserSession(request);
-  const url = new URL(request.url);
-  const state = url.searchParams.get("state");
-  const storedState = session.get("state");
-
-  // 1. Verify that the state matches the one we stored in the session
-  if (state !== storedState) return false;
-
-  // 2. Verify that the hmac is signed
-  const searchParams = new URLSearchParams();
-  for (const [key, value] of url.searchParams) {
-    if (key !== "hmac") searchParams.append(key, value);
-  }
-  const localHmac = crypto
-    .createHmac("sha256", process.env.API_SECRET_KEY!)
-    .update(searchParams.toString())
-    .digest("hex");
-
-  const hmac = url.searchParams.get("hmac");
-  if (localHmac !== hmac) return false;
-  return true;
 }
